Final review comparing with Udacity's project rubric
