+++ app/src/main/java/com/udacity/shoestore/MainActivity.kt
@@ -1,9 +1,6 @@
 package com.udacity.shoestore
 
 import android.os.Bundle
-import android.view.Menu
-import android.view.MenuInflater
-import android.view.MenuItem
 import androidx.appcompat.app.AppCompatActivity
 import androidx.navigation.NavController
 import androidx.navigation.findNavController
@@ -33,20 +30,6 @@ class MainActivity : AppCompatActivity() {
     override fun onSupportNavigateUp(): Boolean {
         val navController = this.findNavController(R.id.nav_host_fragment)
         return NavigationUI.navigateUp(navController, appBarConfiguration)
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
-        navController = this.findNavController(R.id.nav_host_fragment)
-        val inflater = MenuInflater(applicationContext)
-        inflater.inflate(R.menu.menu_logout, menu)
-        return super.onCreateOptionsMenu(menu)
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        when(item.itemId){
-            R.id.item_logout -> navController.navigate(navController.graph.startDestination)
-        }
-        return super.onOptionsItemSelected(item)
     }
 }
