+++ app/src/main/java/com/udacity/shoestore/view/shoeDetails/ShoeDetailsAddFragment.kt
@@ -1,7 +1,10 @@
 package com.udacity.shoestore.view.shoeDetails
 
 import android.os.Bundle
-import android.view.*
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.navigation.fragment.findNavController
@@ -15,16 +18,11 @@ class ShoeDetailsAddFragment : Fragment() {
     private lateinit var binding : FragmentShoeDetailsBinding
     private val viewModel: ShoeListViewModel by activityViewModels()
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setHasOptionsMenu(true)
-    }
-
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentShoeDetailsBinding.inflate(inflater, container, false)
+        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_shoe_details, container, false)
         binding.shoeDetails = Shoe()
 
         binding.btnSave.setOnClickListener {
@@ -36,10 +34,5 @@ class ShoeDetailsAddFragment : Fragment() {
         binding.btnCancel.setOnClickListener { findNavController().navigateUp() }
 
         return binding.root
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        super.onCreateOptionsMenu(menu, inflater)
-        menu.findItem(R.id.item_logout).isVisible = false
     }
 }
