+++ app/src/main/java/com/udacity/shoestore/view/shoeDetails/ShoeDetailsEditFragment.kt
@@ -1,7 +1,10 @@
 package com.udacity.shoestore.view.shoeDetails
 
 import android.os.Bundle
-import android.view.*
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.Observer
@@ -16,16 +19,11 @@ class ShoeDetailsEditFragment : Fragment() {
     private lateinit var binding : FragmentShoeDetailsBinding
     private val viewModel: ShoeListViewModel by activityViewModels()
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setHasOptionsMenu(true)
-    }
-
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentShoeDetailsBinding.inflate(inflater, container, false)
+        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_shoe_details, container, false)
         val shoeDetailsEditFragmentArgs by navArgs<ShoeDetailsEditFragmentArgs>()
         val positionId = shoeDetailsEditFragmentArgs.positionId
 
@@ -41,10 +39,5 @@ class ShoeDetailsEditFragment : Fragment() {
         binding.btnCancel.setOnClickListener { findNavController().navigateUp() }
 
         return binding.root
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        super.onCreateOptionsMenu(menu, inflater)
-        menu.findItem(R.id.item_logout).isVisible = false
     }
 }
