+++ app/src/main/java/com/udacity/shoestore/view/welcome/WelcomeFragment.kt
@@ -4,10 +4,12 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
+import com.udacity.shoestore.R
 import com.udacity.shoestore.databinding.FragmentWelcomeBinding
 
 class WelcomeFragment : Fragment() {
@@ -25,7 +27,7 @@ class WelcomeFragment : Fragment() {
         viewModelFactory = WelcomeViewModelFactory(welcomeFragmentArgs.userName)
         viewModel = ViewModelProvider(this, viewModelFactory).get(WelcomeViewModel::class.java)
 
-        binding = FragmentWelcomeBinding.inflate(inflater, container, false)
+        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_welcome, container, false)
         binding.viewModel = viewModel
         binding.lifecycleOwner = this
