+++ app/src/main/res/navigation/my_navigation.xml
@@ -52,6 +52,9 @@
         <action
             android:id="@+id/action_shoeListFragment_to_shoeDetailsEditFragment"
             app:destination="@id/shoeDetailsEditFragment" />
+        <action
+            android:id="@+id/action_shoeListFragment_to_loginFragment"
+            app:destination="@id/loginFragment" />
     </fragment>
     <fragment
         android:id="@+id/shoeDetailsAddFragment"
